+++ packages/docs/stories/avdate.stories.js
@@ -9,8 +9,8 @@ import AvDate, {
   AvDateRangeField,
 } from '@availity/reactstrap-validation-date';
 import README from '@availity/reactstrap-validation-date/README.md';
-import AvFormResults from './mocks/AvFormResults';
 import '@availity/reactstrap-validation-date/styles.scss';
+import AvFormResults from './mocks/AvFormResults';
 
 const types = {
   text: 'text',
@@ -138,6 +138,7 @@ storiesOf('Components|AvDate', module)
           required={boolean('Required', false)}
           disabled={boolean('Disabled', false)}
           distance={distance}
+          ranges={boolean('ranges', false)}
           errorMessage={text('Generic Error Message', 'This field is invalid')}
           validate={{
             required: {
@@ -194,6 +195,7 @@ storiesOf('Components|AvDate', module)
           min={text('Min Date (yyyy-mm-dd)')}
           max={text('Max Date (yyyy-mm-dd)')}
           distance={distance}
+          ranges={boolean('ranges', false)}
           required={boolean('Required', false)}
           disabled={boolean('Disabled', false)}
           errorMessage={text('Generic Error Message', 'This field is invalid')}